+++ lib/src/di/app_dependencies.dart
@@ -15,6 +15,7 @@ import 'package:where_to_go_today/src/features/authservices/repository/auth_repo
 import 'package:where_to_go_today/src/features/location/location_service.dart';
 import 'package:where_to_go_today/src/features/main/places/service/places_bloc.dart';
 import 'package:where_to_go_today/src/features/main/places/service/repository/places_repository.dart';
+import 'package:where_to_go_today/src/features/main/profile/services/api/profile_api.dart';
 import 'package:where_to_go_today/src/features/main/profile/services/profile_bloc.dart';
 import 'package:where_to_go_today/src/features/main/profile/services/repository/profile_repository.dart';
 import 'package:where_to_go_today/src/features/onboard/services/onboarding_bloc.dart';
@@ -51,7 +52,8 @@ class AppDependencies extends DependencyBundle {
     placesRepository,
   );
 
-  late final profileRepository = ProfileRepository();
+  late final profileApi = ProfileApi(dio);
+  late final profileRepository = ProfileRepository(profileApi);
   late final profileBloc =
       ProfileBloc(profileRepository, authRepository, tokenStorage);