+++ lib/src/features/main/profile/services/api/profile_api.g.dart
@@ -1,13 +1,13 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 
-part of 'places_api.dart';
+part of 'profile_api.dart';
 
 // **************************************************************************
 // RetrofitGenerator
 // **************************************************************************
 
-class _PlacesApi implements PlacesApi {
-  _PlacesApi(this._dio, {this.baseUrl}) {
+class _ProfileApi implements ProfileApi {
+  _ProfileApi(this._dio, {this.baseUrl}) {
     baseUrl ??= 'https://stoplight.io/mocks/softech/wtgt/75539';
   }
 
@@ -16,52 +16,60 @@ class _PlacesApi implements PlacesApi {
   String? baseUrl;
 
   @override
-  Future<PlaceListResponse> fetchPlaces({search, lat, lon, isFavourite}) async {
-    const _extra = <String, dynamic>{};
-    final queryParameters = <String, dynamic>{
-      r'search': search,
-      r'lat': lat,
-      r'lon': lon,
-      r'fav': isFavourite
-    };
-    queryParameters.removeWhere((k, v) => v == null);
-    final _headers = <String, dynamic>{};
-    final _data = <String, dynamic>{};
-    final _result = await _dio.fetch<Map<String, dynamic>>(
-        _setStreamType<PlaceListResponse>(
-            Options(method: 'GET', headers: _headers, extra: _extra)
-                .compose(_dio.options, '/places',
-                    queryParameters: queryParameters, data: _data)
-                .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
-    final value = PlaceListResponse.fromJson(_result.data!);
-    return value;
-  }
-
-  @override
-  Future<Place> fetchPlace(id) async {
+  Future<UserProfile> getUserProfile() async {
     const _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     final _headers = <String, dynamic>{};
     final _data = <String, dynamic>{};
     final _result = await _dio.fetch<Map<String, dynamic>>(
-        _setStreamType<Place>(
+        _setStreamType<UserProfile>(
             Options(method: 'GET', headers: _headers, extra: _extra)
-                .compose(_dio.options, '/place/${id}',
+                .compose(_dio.options, '/user',
                     queryParameters: queryParameters, data: _data)
                 .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
-    final value = Place.fromJson(_result.data!);
+    final value = UserProfile.fromJson(_result.data!);
     return value;
   }
 
   @override
-  Future<void> setFavourite(id, body) async {
+  Future<void> changeProfile(body) async {
     const _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     final _headers = <String, dynamic>{};
     final _data = body;
     await _dio.fetch<void>(_setStreamType<void>(
         Options(method: 'PUT', headers: _headers, extra: _extra)
-            .compose(_dio.options, '/place/${id}/favourite',
+            .compose(_dio.options, '/user',
+                queryParameters: queryParameters, data: _data)
+            .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
+    return null;
+  }
+
+  @override
+  Future<UserSettings> getUserSettings() async {
+    const _extra = <String, dynamic>{};
+    final queryParameters = <String, dynamic>{};
+    final _headers = <String, dynamic>{};
+    final _data = <String, dynamic>{};
+    final _result = await _dio.fetch<Map<String, dynamic>>(
+        _setStreamType<UserSettings>(
+            Options(method: 'GET', headers: _headers, extra: _extra)
+                .compose(_dio.options, '/user/settings',
+                    queryParameters: queryParameters, data: _data)
+                .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
+    final value = UserSettings.fromJson(_result.data!);
+    return value;
+  }
+
+  @override
+  Future<void> changeUserSettings(body) async {
+    const _extra = <String, dynamic>{};
+    final queryParameters = <String, dynamic>{};
+    final _headers = <String, dynamic>{};
+    final _data = body;
+    await _dio.fetch<void>(_setStreamType<void>(
+        Options(method: 'PUT', headers: _headers, extra: _extra)
+            .compose(_dio.options, '/user/settings',
                 queryParameters: queryParameters, data: _data)
             .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
     return null;