+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+- feat: Places Screen
 - feat: Place Card
 - feat: CodeVm - add authorization via OTP
 - feat: SignInScreenVm - add phone authorization
@@ -7,6 +8,7 @@
 - feat: MainScreen + PlacesScreen
 - feat: OnboardStorage
 - feat: profile screen
+- feat: profile Api
 - feat: places api
 - feat: profile slivers
 - feat: profile bloc